# docs/swagger.yaml
basePath: /api/v1
definitions:
  bengkel_internal_http_response.APIError:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/bengkel_internal_http_response.FieldError'
        type: array
      message:
        type: string
    type: object
  bengkel_internal_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/bengkel_internal_http_response.APIError'
      meta:
        $ref: '#/definitions/bengkel_internal_http_response.Meta'
    type: object
  bengkel_internal_http_response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
  bengkel_internal_http_response.Meta:
    properties:
      last_page:
        type: integer
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
...
  internal_http_handler.intakeCustomerInput:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        maxLength: 150
        type: string
      phone:
        maxLength: 30
        type: string
    required:
    - name
    type: object
  internal_http_handler.intakeInput:
    properties:
      complaint:
        maxLength: 1000
        type: string
      customer:
        $ref: '#/definitions/internal_http_handler.intakeCustomerInput'
      customer_id:
        type: string
      mechanic_id:
        type: string
      odometer:
        minimum: 0
        type: integer
      vehicle:
        $ref: '#/definitions/internal_http_handler.intakeVehicleInput'
      vehicle_id:
        type: string
    required:
    - complaint
    type: object
  internal_http_handler.intakeVehicleInput:
    properties:
      brand:
        maxLength: 80
        type: string
      color:
        maxLength: 50
        type: string
      identifier:
        maxLength: 100
        type: string
      model:
        maxLength: 100
        type: string
      plate_number:
        maxLength: 30
        type: string
      year:
        minimum: 0
        type: integer
    required:
    - identifier
    type: object
...
  internal_http_handler.productInput:
    properties:
      cost_price:
        minimum: 0
        type: integer
      is_active:
        type: boolean
      min_stock:
        minimum: 0
        type: number
      name:
        maxLength: 180
        type: string
      selling_price:
        minimum: 0
        type: integer
      sku:
        maxLength: 80
        type: string
      type:
        enum:
        - part
        - service
        - other
        type: string
      unit:
        maxLength: 30
        type: string
    required:
    - name
    - sku
    - type
    - unit
    type: object
...
  internal_http_handler.vehicleInput:
    properties:
      brand:
        maxLength: 80
        type: string
      color:
        type: string
      customer_id:
        type: string
      identifier:
        maxLength: 100
        type: string
      model:
        maxLength: 100
        type: string
      notes:
        type: string
      odometer:
        minimum: 0
        type: integer
      plate_number:
        maxLength: 30
        type: string
      year:
        minimum: 0
        type: integer
    required:
    - identifier
    type: object
...
  internal_http_handler.workOrderUpdateInput:
    properties:
      diagnosis:
        maxLength: 3000
        type: string
      mechanic_id:
        type: string
      status:
        enum:
        - inspection
        - approval
        - in_progress
        - completed
        - cancelled
        type: string
    required:
    - status
    type: object
info:
  contact: {}
  description: API operasional bengkel, kasir, stok, pembayaran, akuntansi, audit,
    dan CMS.
  title: BengkelOS API
  version: "1.0"
paths:
...
  /inventory/movements:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Riwayat pergerakan persediaan
      tags:
      - Inventory
...
  /products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Nonaktifkan produk yang sudah pernah dipakai
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Produk
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.productInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Ubah barang atau jasa
      tags:
      - Products
...
  /vehicles/{id}:
    delete:
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Hapus kendaraan yang belum memiliki work order
      tags:
      - Vehicles
    put:
      consumes:
      - application/json
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: string
      - description: Kendaraan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.vehicleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Ubah data kendaraan
      tags:
      - Vehicles
...
  /work-orders/{id}:
    get:
      parameters:
      - description: Work order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Detail pekerjaan, motor, pelanggan, dan item terpakai
      tags:
      - Work Orders
    patch:
      consumes:
      - application/json
      parameters:
      - description: Work order ID
        in: path
        name: id
        required: true
        type: string
      - description: Pekerjaan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.workOrderUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Perbarui diagnosis, montir, dan status pekerjaan
      tags:
      - Work Orders
...
  /work-orders/{id}/items/{item_id}:
    delete:
      parameters:
      - description: Work order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Hapus item dan kembalikan stok barang
      tags:
      - Work Orders
  /work-orders/intake:
    post:
      consumes:
      - application/json
      parameters:
      - description: Penerimaan motor
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.intakeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Terima motor baru atau kendaraan terdaftar
      tags:
      - Work Orders
  /work-orders/options/mechanics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Daftar montir aktif pada cabang
      tags:
      - Work Orders
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Masukkan "Bearer {access_token}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"